Add PodSecurityPolicy Support
Add support for PodSecurityPolicy's, allowing us to disable use of the
hostPath volume type.

This change adds 3 PSP's:

* lockdown

The lockdown PodSecurityPolicy is intended to be as locked
down as is possible. It's likely to be so restrictive as to not generally
be usable.

* unprivileged (Default assigned to all users)

The unprivileged PodSecurityPolicy is intended to be a
reasonable compromise between suse:caasp:psp:lockdown, and
suse:caasp:psp:privileged. By default, we'll grant this PSP to all
users and service accounts.

* privileged

The privileged PodSecurityPolicy is intended to be given
only to trusted workloads. It provides for as few restrictions as possible
and should only be assigned to highly trusted users.

Fixes bsc#1047535